requirements.txt
python-telegram-bot==20.7
openai==1.3.7
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.0